+++ resources/js/components/System/UserItem.vue
@@ -1,5 +1,5 @@
 <template>
-    <sui-list-item>
+    <sui-list-item :active="active">
         <sui-list-icon name="user" size="large" vertical-align="middle" />
         <sui-list-content>
             <a is="sui-list-header" @click="edit">{{ name }}</a>
@@ -12,6 +12,7 @@
 export default {
     props: {
         user: Object,
+        active: Boolean,
     },
     data () {
         return {