+++ resources/js/pages/Databases.vue
@@ -1,13 +1,48 @@
 <template>
     <sui-grid container>
         <sui-grid-column>
-            <sui-segment class="placeholder">
-                <sui-header icon="bug">
-                    Oops!<br><small>
-                        This section hasn't been made yet&hellip;
-                    </small>
-                </sui-header>
-            </sui-segment>
+
+            <sui-input placeholder="Type a name for your database..."
+                    icon="search" class="fluid massive"
+                    :inverted="darkMode" :transparent="darkMode"
+                    v-model="search" @keydown.enter="create"></sui-input>
+
+            <sui-table selectable :inverted="darkMode">
+                <sui-table-row v-for="db, key in databases" :key="key">
+                    <sui-table-cell>
+                        <sui-icon :color="darkMode ? 'orange' : 'violet'" name="database" /> {{ db }}
+                    </sui-table-cell>
+                </sui-table-row>
+            </sui-table>
+
         </sui-grid-column>
     </sui-grid>
 </template>
+
+<script>
+import { mapActions, mapGetters } from 'vuex';
+
+export default {
+    mounted () {
+        this.$store.dispatch('databases/load');
+    },
+    computed: {
+        ...mapGetters({
+            databases: 'databases/filtered',
+        }),
+        search: {
+            get () {
+                return this.$store.state.databases.search;
+            },
+            set (value) {
+                this.$store.dispatch('databases/filter', value);
+            },
+        },
+    },
+    methods: {
+        ...mapActions({
+            create: 'databases/create',
+        }),
+    },
+};
+</script>