+++ resources/js/pages/Sites.vue
@@ -2,9 +2,9 @@
     <sui-grid container>
         <sui-grid-row>
             <sui-grid-column>
-                <sui-input placeholder="Type a name for your Application..."
-                        :icon="filterIcon" class="fluid massive"
-                        v-model="site.name" @input="filterSites" @keyup.enter="createOrEdit"></sui-input>
+                <sui-input placeholder="Type a name for your Application..." :icon="filterIcon"
+                           class="fluid massive" :inverted="darkMode" :transparent="darkMode"
+                           v-model="site.name" @input="filterSites" @keyup.enter="createOrEdit"></sui-input>
             </sui-grid-column>
         </sui-grid-row>
         <router-view id="sites" :sites="sites" />
@@ -57,7 +57,9 @@ export default {
                 return this.$router.push({ name: 'apps.edit', params: { id: result.id }});
             }
 
-            this.$store.dispatch('sites/create');
+            this.$store.dispatch('sites/create').then(({data})=>{
+                this.$router.push({ name: 'apps.edit', params: {id: data.id}})
+            });
         },
     }
 }