+++ resources/js/pages/Sites/Detail.vue
@@ -6,7 +6,7 @@
             </sui-card-group>
         </sui-grid-column>
         <sui-grid-column :width=12>
-            <h2 is="sui-header" size="huge">
+            <h2 is="sui-header" size="huge" :inverted="darkMode">
                 <router-link :to="{ name: 'apps.edit', params: { id: site.id } }"
                              content="Manage Site" is="sui-button" floated="right"
                              color="teal" icon="cogs" size="large" />
@@ -16,12 +16,13 @@
                 </sui-header-subheader>
             </h2>
 
-            <sui-header attached="top">Source Files</sui-header>
-            <sui-segment attached>
+            <sui-header attached="top" :inverted="darkMode">Source Files</sui-header>
+            <sui-segment attached :inverted="darkMode">
                 <sui-grid>
                     <sui-grid-row :columns="2">
                         <sui-grid-column>
-                            <sui-header size="tiny"> Repository Clone URL
+                            <sui-header size="tiny" :inverted="darkMode">
+                                Repository Clone URL
                                 <sui-header-subheader v-if="site.source_repo">
                                     {{ site.source_repo }}
                                 </sui-header-subheader>
@@ -31,7 +32,7 @@
                             </sui-header>
                         </sui-grid-column>
                         <sui-grid-column>
-                            <sui-header size="tiny"> Tracking Branch
+                            <sui-header size="tiny" :inverted="darkMode"> Tracking Branch
                                 <sui-button basic positive icon="download" floated="right"
                                     content="Pull Latest Code" @click="pullFiles(site)" />
                                 <sui-header-subheader v-if="site.source_branch">
@@ -45,7 +46,7 @@
                     </sui-grid-row>
                 </sui-grid>
 
-                <sui-header size="tiny" v-if="site.document_root">
+                <sui-header size="tiny" :inverted="darkMode" v-if="site.document_root">
                     <router-link :to="{ name: 'files', params: {path: site.document_root } }"
                                  content="Browse files" is="sui-button" floated="right"
                                  basic primary icon="open folder" />
@@ -55,17 +56,23 @@
                 <p v-else>This project doesn't have a document root defined.</p>
             </sui-segment>
 
-            <sui-header v-if="site.type == 'redirect'" attached="top">Redirection</sui-header>
-            <sui-segment v-if="site.type == 'redirect'" attached>
+            <sui-header v-if="site.type == 'redirect'" attached="top" :inverted="darkMode">
+                Redirection
+            </sui-header>
+            <sui-segment v-if="site.type == 'redirect'" attached :inverted="darkMode">
                 <sui-grid>
                     <sui-grid-row :columns="2">
                         <sui-grid-column>
-                            <sui-header size="tiny"> Target URL
-                                <sui-header-subheader>{{ site.redirect_to }}</sui-header-subheader>
+                            <sui-header size="tiny" :inverted="darkMode">
+                                Target URL
+                                <sui-header-subheader>
+                                    {{ site.redirect_to }}
+                                </sui-header-subheader>
                             </sui-header>
                         </sui-grid-column>
                         <sui-grid-column>
-                            <sui-header size="tiny"> Redirect Type
+                            <sui-header size="tiny" :inverted="darkMode">
+                                Redirect Type
                                 <sui-header-subheader v-if="site.redirect_type == 301">
                                     Permanent
                                 </sui-header-subheader>